simplelink_lpf3: ti: drivers: power: fix fault after mcuboot jump to app
When using with MCUBoot, the system is already  initialized before jumping
to the application. When the HFXT is started with `PowerCC23X0_startHFXT()`
the `PowerCC23X0_oscillatorISR()` gets triggered, which cancels the timer
and sets the callback to NULL before `ClockP_start()` is called. As a
result, ClockP dereferences a NULL pointer when the timer expires and then
we hit a fault. To fix, init of HFXT and HFXT timeout clock are put in a
critical section.

Signed-off-by: Ging Gonzalez <g-gonzalez@ti.com>