net: slnetsock: do not require posix to define SlNetSock_Timeval_t
Previously, the header `<ti/net/slnetsock.h>` was forward-declaring a
typed struct when deprecated and nonexistent Kconfig options were not
present. This caused a compile error when building
`sample.net.sockets.http_get.offload.simplelink` because struct timeval
was not defined in a non-standard header.

Since the offload driver must exist whether or not the applciation
has enabled POSIX, and struct timeval is POSIX, simply define the
SlNetSock_Timeval_t typed structure whether or not POSIX is enabled using
ANSI/ISO C types.

In this case, `long` was chosen to match the wordsize of whatever
architecture is being built.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>