simplelink: wff3: fix macro redefinitions
Zephyr already provides definitions for the following macros:
* __time_t_defined
* ARRAY_SIZE
* USEC_PER_SEC

USEC_PER_SEC is not used by the Wi-Fi stack and can be removed.
The remaining macros can be wrapped in #if !defined().

MAC_ADDR_LEN is defined both in:
* source/ti/drivers/net/wifi/wifi_host_driver/trnspt_layer/ethernet.h
* source/ti/net/wifi_stack/driver/drv_ti/802_11defs.h
but ethernet.h also provides a definition for MAC_ADDR_SIZE, which seems
to be used in the transport layer context in lieu of MAC_ADDR_LEN.
As such, drop the duplicate definition from ethernet.h.

Signed-off-by: Artur Rojek <artur@conclusive.pl>